search_api/requirements.txt
httpx[socks]>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
orjson>=3.11.6
pydantic>=2.0
python-dotenv>=1.0.0
